MASSIVE REFACTORING with Errors
