+++ CMakeLists.txt
@@ -2,6 +2,10 @@ cmake_minimum_required(VERSION 3.13 FATAL_ERROR)
 
 project(to)
 
+set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+
+
 add_subdirectory(thirdparty)
 
 include_directories(${CMAKE_SOURCE_DIR}/rapidjson
