+++ to/CMakeLists.txt
@@ -1,22 +1,24 @@
 cmake_minimum_required(VERSION 3.13 FATAL_ERROR)
 
-set(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
 project(exchange)
 
+add_subdirectory(models)
 add_subdirectory(mapper)
 add_subdirectory(optimizer)
 add_subdirectory(logger)
+add_subdirectory(connectors)
+add_subdirectory(controlers)
 
-add_executable(${PROJECT_NAME} exchange.cpp logger/logger.cpp logger/include/logger.h) #cmake is a stupid shit
+add_executable(${PROJECT_NAME} exchange.cpp) #cmake is a stupid shit
 
 set_target_properties(${PROJECT_NAME} PROPERTIES
         RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}")
 
-target_link_libraries(${PROJECT_NAME}
-                      "${TORCH_LIBRARIES}"
+target_link_libraries(${PROJECT_NAME} PUBLIC
                       stdc++fs
+                      models
                       logger
+                      controlers
+                      connectors
                       optimizer
                       mapper)
