+++ to/logger/include/logger.h
@@ -2,8 +2,8 @@
 // Created by Frédéric Gardes<frederic.gardes@orange.com> on 5/27/19.
 //
 
-#ifndef COMMUNICATION_LAYER_LOGGER_H
-#define COMMUNICATION_LAYER_LOGGER_H
+#ifndef TO_LOGGER_H
+#define TO_LOGGER_H
 
 #include <string>
 #include <fstream>
@@ -13,4 +13,4 @@ namespace logger {
     auto write(const std::string &text) -> void;
 }
 
-#endif //COMMUNICATION_LAYER_LOGGER_H
+#endif //TO_LOGGER_H
