+++ to/models/maneuver_recommendation.cpp
@@ -4,169 +4,7 @@
 // Created by: KCL
 
 // Modified by: Omar Nassef (KCL)
-
-#include <iostream>
-#include <ostream>
-#include <string>
-#include <vector>
-#include <tuple>
-#include "waypoint.cpp"
-
-using std::string;
-using std::vector;
-using std::ostream;
-using std::tuple;
-
-typedef uint32_t uint4;
-
-class ManeuverRecommendation {
-
-private:
-
-    string type;
-    string context;
-    string origin;
-    string version = "1.2.0";
-    uint64_t timestamp;
-    string uuid_vehicle; // Digital Identifier of RU that shall execute the maneuver.
-    string uuid_to; // Digital Identifier of the traffic orchestrator.
-    string uuid_maneuver; // Digital Identifier of the maneuver recommendation, generated by traffic orchestrator.
-    vector<std::shared_ptr<Waypoint>> waypoints; // A list of waypoints indicating a target position, speed, and lane for the vehicle.
-
-    /* The vehicle should start adapting towards the target lane and speed at the indicated timestamp and position. */
-    uint64_t timestamp_action;
-    uint32_t latitude_action;
-    uint32_t longitude_action;
-    uint16_t speed_action;
-    uint4 lane_position;
-    string source_uuid;
-    string uuid;
-    string message_id;
-    string signature;
-
-public:
-
-    ManeuverRecommendation(tuple<uint8_t, uint8_t, uint8_t> version,
-                           uint64_t timestamp,
-                           string uuid_vehicle,
-                           string uuid_to,
-                           string uuid_maneuver,
-                           vector<std::shared_ptr<Waypoint>> waypoints,
-                           uint64_t timestamp_action,
-                           uint32_t latitude_action,
-                           uint32_t longitude_action,
-                           uint16_t speed_action,
-                           uint4 lane_position,
-                           string signature) :
-            type(type),
-            context(context),
-            origin(origin),
-            timestamp(timestamp),
-            uuid_vehicle(uuid_vehicle),
-            uuid_to(uuid_to),
-            uuid_maneuver(uuid_maneuver),
-            waypoints(waypoints),
-            timestamp_action(timestamp_action),
-            latitude_action(latitude_action),
-            longitude_action(longitude_action),
-            speed_action(speed_action),
-            lane_position(lane_position),
-            signature(signature) {
-        type = "maneuver";
-        context = "lane_merge";
-        origin = "traffic_orchestrator";
-        source_uuid = "traffic_orchestrator";
-        uuid_maneuver = to_string(10000000 + (std::rand() % (99999999 - 10000000 + 1)));
-
-    }
-
-    ManeuverRecommendation() {
-        type = "maneuver";
-        context = "lane_merge";
-        origin = "traffic_orchestrator";
-        source_uuid = "traffic_orchestrator";
-        uuid_maneuver = to_string(10000000 + (std::rand() % (99999999 - 10000000 + 1)));
-    }
-
-
-    friend std::ostream &operator<<(ostream &os, ManeuverRecommendation *maneuverRec); // Overload << to print a recommendation.
-
-    string getType();
-
-    string getContext();
-
-    string getOrigin();
-
-    string getVersion();
-
-    uint64_t getTimestamp();
-
-    string getUuidVehicle();
-
-    string getUuidTo();
-
-    string getUuidManeuver();
-
-    auto getWaypoints();
-
-    uint64_t getTimestampAction();
-
-    uint32_t getLatitudeAction();
-
-    uint32_t getLongitudeAction();
-
-    uint16_t getSpeedAction();
-
-    uint4 getLanePositionAction();
-
-    string getSignature();
-
-    string getSourceUUID();
-
-    string getUUID();
-
-    string getMessageID();
-
-    void setType(string);
-
-    void setContext(string);
-
-    void setOrigin(string);
-
-    void setVersion(string);
-
-    void setTimestamp(uint64_t);
-
-    void setUuidVehicle(string);
-
-    void setUuidTo(string);
-
-    void setUuidManeuver(string);
-
-    void setWaypoints(vector<std::shared_ptr<Waypoint>> waypoints);
-
-    void setTimestampAction(uint64_t);
-
-    void setLatitudeAction(uint32_t);
-
-    void setLongitudeAction(uint32_t);
-
-    void setSpeedAction(uint16_t);
-
-    void setLanePositionAction(uint4);
-
-    void setSignature(string);
-
-    void setSourceUUID(string);
-
-    void setUUID(string);
-
-    void setMessageID(string);
-
-    void addWaypoint(const std::shared_ptr<Waypoint> &waypoint);
-
-    void emptyWaypoints();
-};
+#include "include/maneuver_recommendation.h"
 
 string ManeuverRecommendation::getType() { return type; }
 
@@ -184,7 +22,7 @@ string ManeuverRecommendation::getUuidTo() { return uuid_to; }
 
 string ManeuverRecommendation::getUuidManeuver() { return uuid_maneuver; }
 
-auto ManeuverRecommendation::getWaypoints() { return waypoints; }
+vector<std::shared_ptr<Waypoint>> ManeuverRecommendation::getWaypoints() { return waypoints; }
 
 uint64_t ManeuverRecommendation::getTimestampAction() { return timestamp_action; }
