--- to/optimizer/include/optimizer-engine.h
@@ -1,17 +1,0 @@
-//
-// Created by jab on 27/05/19.
-//
-
-#ifndef COMMUNICATION_LAYER_ENGINE_H
-#define COMMUNICATION_LAYER_ENGINE_H
-
-
-class OptimizerEngine {
-public:
-
-    OptimizerEngine();
-
-};
-
-
-#endif //COMMUNICATION_LAYER_ENGINE_H
