+++ to/optimizer/include/optimizer_engine.h
@@ -1,0 +1,17 @@
+//
+// Created by jab on 27/05/19.
+//
+
+#ifndef TO_OPTIMIZER_ENGINE_H
+#define TO_OPTIMIZER_ENGINE_H
+
+
+class OptimizerEngine {
+public:
+
+    OptimizerEngine();
+
+};
+
+
+#endif //TO_ENGINE_H
